requirements/requirements-base.txt
backoff==2.2.1
beautifulsoup4==4.12.3 
clickhouse-connect==0.6.6
coloredlogs==15.0.1
duckdb==0.8.1
fastapi==0.115.13
flask==3.0.3
flask-login==0.6.3
flask-session==0.8.0
flask-cors==5.0.0
authlib==1.3.0
flatbuffers==23.5.26
gitpython==3.1.45
hnswlib==0.7.0
httptools==0.6.1
httpx==0.27.2
humanfriendly==10.0
croniter==2.0.5
langgraph==1.0.1
langchain-mcp-adapters==0.1.11
langchain==1.0.2
langchain_ollama==1.0.0
loguru==0.7.2
lz4==4.3.2
mistune==3.0.2
monotonic==1.6
mkdocs==1.6.1
mkdocs-material
overrides==7.4.0
piazza-api==0.14.0
posthog==3.3.1
pygments==2.18.0
pypdf==3.16.1
python-dotenv==1.0.0
python-redmine==2.4.0
regex==2023.12.25
sentencepiece==0.2.0
sympy==1.13.1
tiktoken==0.7.0
uvloop==0.19.0
watchfiles==0.21.0
zstandard==0.25.0
langchain-openai==1.0.1
langchain-community==0.4
langchain-core==1.0.0
langchain-text-splitters==1.0.0
langchain-huggingface==1.0.0
langchain-anthropic==1.0.0
langchain-classic==1.0.0
unstructured==0.18.2
openai==2.2.0
pydantic==2.11.7
pydantic-core==2.33.2
anthropic==0.71.0
scipy==1.13.1
urllib3==2.3.0
pulsar-client==3.5.0
psycopg2-binary==2.9.10
requests==2.32.5
pyyaml==6.0.2
protobuf==4.25.8
jira==3.8.0
spacy==3.8.7
qwen-vl-utils==0.0.11
rapidfuzz==3.13.0
selenium==4.32.0
multidict==6.6.2
murmurhash==1.0.12
numpy==1.26.4
importlib_metadata==7.2.1
marshmallow==3.26.0
llguidance==0.7.27
jiter==0.9.1
idna==3.10
httpcore==1.0.7
cymem==2.0.10
dataclasses_json==0.6.7
distro==1.9.0
charset_normalizer==3.4.2
anyio==4.8.0
catalogue==2.0.9
aiohttp==3.9.5
nltk==3.9.1
sentence-transformers==5.1.2
rank_bm25==0.2.2
